refactor: fixing typo
